requirements.txt
gitdb==0.6.4
GitPython==1.0.2
smmap==0.9.0
docopt==0.6.2